+++ dubbo-remoting/dubbo-remoting-netty4/src/test/java/org/apache/dubbo/remoting/transport/netty4/ReplierDispatcherTest.java
@@ -14,92 +14,122 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.dubbo.remoting;
+package org.apache.dubbo.remoting.transport.netty4;
 
 import org.apache.dubbo.common.URL;
+import org.apache.dubbo.remoting.RemotingException;
 import org.apache.dubbo.remoting.exchange.ExchangeChannel;
+import org.apache.dubbo.remoting.exchange.ExchangeServer;
 import org.apache.dubbo.remoting.exchange.Exchangers;
 import org.apache.dubbo.remoting.exchange.support.Replier;
 import org.apache.dubbo.remoting.exchange.support.ReplierDispatcher;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 import java.io.Serializable;
 import java.util.Random;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.Executor;
-import java.util.concurrent.Executors;
+import java.util.concurrent.*;
+
 
 /**
- * Main
+ * ReplierDispatcherTest
  */
 
-public class Main {
-    public static void main(String[] args) throws Exception {
-        startServer(9010);
-        mutliThreadTest(10, 9010);
-        dataPackageTest(9010);
-    }
+public class ReplierDispatcherTest {
 
-    private static void startServer(int port) throws Exception {
+    private ExchangeServer exchangeServer;
+
+    private ConcurrentHashMap<String, ExchangeChannel> clients = new ConcurrentHashMap<>();
+
+    private int port;
+
+    @BeforeEach
+    void startServer() throws RemotingException {
+        port = (int) (1000 * Math.random() + 10000);
         ReplierDispatcher dispatcher = new ReplierDispatcher();
         dispatcher.addReplier(RpcMessage.class, new RpcMessageHandler());
         dispatcher.addReplier(Object.class, new Replier<Object>() {
             public Object reply(ExchangeChannel channel, Object msg) {
-                for (int i = 0; i < 10000; i++)
+                for (int i = 0; i < 10000; i++) {
                     System.currentTimeMillis();
+                }
                 System.out.println("handle:" + msg + ";thread:" + Thread.currentThread().getName());
                 return new StringMessage("hello world");
             }
         });
-        Exchangers.bind(URL.valueOf("dubbo://localhost:" + port), dispatcher);
+        exchangeServer = Exchangers.bind(URL.valueOf("dubbo://localhost:" + port), dispatcher);
     }
 
-    static void dataPackageTest(int port) throws Exception {
+
+    @Test
+    public void testDataPackage() throws Exception {
         ExchangeChannel client = Exchangers.connect(URL.valueOf("dubbo://localhost:" + port));
         Random random = new Random();
         for (int i = 5; i < 100; i++) {
             StringBuilder sb = new StringBuilder();
             for (int j = 0; j < i * 100; j++)
                 sb.append("(" + random.nextLong() + ")");
-            Main.Data d = new Main.Data();
+            Data d = new Data();
             d.setData(sb.toString());
-            client.request(d).get();
+            Assertions.assertEquals(client.request(d).get().toString(), "hello world");
         }
         System.out.println("send finished.");
+        clients.put(Thread.currentThread().getName(), client);
     }
 
-    static void mutliThreadTest(int tc, final int port) throws Exception {
-        Executor exec = Executors.newFixedThreadPool(tc);
+
+    @Test
+    public void testMutliThread() throws Exception {
+        int tc = 10;
+        ExecutorService exec = Executors.newFixedThreadPool(tc);
         for (int i = 0; i < tc; i++)
             exec.execute(new Runnable() {
                 public void run() {
                     try {
-                        test(port);
+                        clientExchangeInfo(port);
                     } catch (Exception e) {
                         e.printStackTrace();
                     }
                 }
             });
+        exec.shutdown();
+        exec.awaitTermination(10, TimeUnit.SECONDS);
     }
 
-    private static void test(int port) throws Exception {
+    void clientExchangeInfo(int port) throws Exception {
         ExchangeChannel client = Exchangers.connect(URL.valueOf("dubbo://localhost:" + port));
+        clients.put(Thread.currentThread().getName(), client);
         MockResult result = (MockResult) client.request(new RpcMessage(DemoService.class.getName(), "plus", new Class<?>[]{int.class, int.class}, new Object[]{55, 25})).get();
-        System.out.println("55+25=" + result.getResult());
-
-        for (int i = 0; i < 100; i++)
-            client.request(new RpcMessage(DemoService.class.getName(), "sayHello", new Class<?>[]{String.class}, new Object[]{"qianlei" + i}));
-
-        for (int i = 0; i < 100; i++)
-            client.request(new Main.Data());
-
-        System.out.println("=====test invoke=====");
+        Assertions.assertEquals(result.getResult(), 80);
         for (int i = 0; i < 100; i++) {
-            CompletableFuture<Object> future = client.request(new Main.Data());
-            System.out.println("invoke and get");
-            System.out.println("invoke result:" + future.get());
+            client.request(new RpcMessage(DemoService.class.getName(), "sayHello", new Class<?>[]{String.class}, new Object[]{"qianlei" + i}));
         }
-        System.out.println("=====the end=====");
+        for (int i = 0; i < 100; i++) {
+            client.request(new Data());
+        }
+        for (int i = 0; i < 100; i++) {
+            CompletableFuture<Object> future = client.request(new Data());
+            Assertions.assertEquals(future.get().toString(), "hello world");
+        }
     }
+
+
+    @AfterEach
+    void tearDown() {
+        try {
+            if (exchangeServer != null)
+                exchangeServer.close();
+        } finally {
+            if (clients.size() != 0)
+                clients.forEach((key, value) -> {
+                    value.close();
+                    clients.remove(key);
+                });
+        }
+    }
+
 
     static class Data implements Serializable {
         private static final long serialVersionUID = -4666580993978548778L;
